Update score submission serialisation tests to use `SubmittableScore`
