+++ osu.Game.Tests/Online/TestAPIModJsonSerialization.cs
@@ -8,6 +8,7 @@ using osu.Framework.Bindables;
 using osu.Game.Beatmaps;
 using osu.Game.Configuration;
 using osu.Game.Online.API;
+using osu.Game.Online.Solo;
 using osu.Game.Rulesets;
 using osu.Game.Rulesets.Difficulty;
 using osu.Game.Rulesets.Mods;
@@ -88,33 +89,27 @@ namespace osu.Game.Tests.Online
         }
 
         [Test]
-        public void TestDeserialiseScoreInfoWithEmptyMods()
+        public void TestDeserialiseSubmittableScoreWithEmptyMods()
         {
-            var score = new ScoreInfo { Ruleset = new OsuRuleset().RulesetInfo };
+            var score = new SubmittableScore(new ScoreInfo { Ruleset = new OsuRuleset().RulesetInfo });
 
-            var deserialised = JsonConvert.DeserializeObject<ScoreInfo>(JsonConvert.SerializeObject(score));
-
-            if (deserialised != null)
-                deserialised.Ruleset = new OsuRuleset().RulesetInfo;
+            var deserialised = JsonConvert.DeserializeObject<SubmittableScore>(JsonConvert.SerializeObject(score));
 
             Assert.That(deserialised?.Mods.Length, Is.Zero);
         }
 
         [Test]
-        public void TestDeserialiseScoreInfoWithCustomModSetting()
+        public void TestDeserialiseSubmittableScoreWithCustomModSetting()
         {
-            var score = new ScoreInfo
+            var score = new SubmittableScore(new ScoreInfo
             {
                 Ruleset = new OsuRuleset().RulesetInfo,
                 Mods = new Mod[] { new OsuModDoubleTime { SpeedChange = { Value = 2 } } }
-            };
+            });
 
-            var deserialised = JsonConvert.DeserializeObject<ScoreInfo>(JsonConvert.SerializeObject(score));
+            var deserialised = JsonConvert.DeserializeObject<SubmittableScore>(JsonConvert.SerializeObject(score));
 
-            if (deserialised != null)
-                deserialised.Ruleset = new OsuRuleset().RulesetInfo;
-
-            Assert.That(((OsuModDoubleTime)deserialised?.Mods[0])?.SpeedChange.Value, Is.EqualTo(2));
+            Assert.That((deserialised?.Mods[0])?.Settings["speed_change"], Is.EqualTo(2));
         }
 
         private class TestRuleset : Ruleset
